requirements.txt
python-telegram-bot==11.1.0
sqlalchemy==1.3.3
psycopg2-binary==2.7.5
persian==0.4.0
